Fix badly cropped sprites
This was a regression introduced by the switch to
sf::Sprite::setTextureRect() to define the section of the texture
that should be used in the sprite. The old format expect the absolute position
of the top-left and bottom-right angles, while the new API expects only the top-left
plus width and height.